Fix version constraints and broken tests
Got rid of minitest-spec-rails which was causing the dependency hell on
this project, and converted all the minitest-spec syntax into regular
old xunit-style syntax for both ease-of-use and to reduce dependencies.
Additionally, removed a "version test" that is totally useless. Should
resolve our testing and versioning issues.

Shoutout to @connorshea for the help on this!